refactor: LottosGenerator 클래스 구현하여 Lottos 생성역할 위임
