+++ src/main/java/lotto/domain/Lottos.java
@@ -1,31 +1,21 @@
 package lotto.domain;
 
-import lotto.utils.RandomLottoNumbersGenerator;
-
-import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
 public class Lottos implements Iterable<Lotto>{
-    List<Lotto> lottoList = new ArrayList<>();
-
-    public Lottos(){
-    }
+    List<Lotto> lottos;
 
     public Lottos(List<Lotto> lottoList){
-        this.lottoList = lottoList;
-    }
-
-    public void addRandomLotto(){
-        lottoList.add(new Lotto(RandomLottoNumbersGenerator.generate()));
+        this.lottos = lottoList;
     }
 
     public int size() {
-        return lottoList.size();
+        return lottos.size();
     }
 
     @Override
     public Iterator<Lotto> iterator() {
-        return lottoList.iterator();
+        return lottos.iterator();
     }
 }
