+++ src/main/java/lotto/generator/RandomLottoNumbersGenerator.java
@@ -1,4 +1,4 @@
-package lotto.utils;
+package lotto.generator;
 
 import lotto.domain.LottoNumber;
