+++ src/test/java/lotto/domain/MoneyTest.java
@@ -1,11 +1,9 @@
 package lotto.domain;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -40,7 +38,7 @@ public class MoneyTest {
         assertThat(money.lottoCountToBuy()).isEqualTo(count);
     }
 
-    @ParameterizedTest(name="수익률 계산")
+    @ParameterizedTest(name="수익률 계산: 초기금액 {0}원 - 상금 {1}원 - 수익률 {2}")
     @CsvSource(value = {"100000:5000:0.05", "100500:5000:0.05"}, delimiter = ':')
     void Money_수익률_계산(int initialMoney, int winningPrice, double profit){
         Money money = new Money(initialMoney);
