+++ src/test/java/lotto/generator/RandomLottoNumbersGeneratorTest.java
@@ -1,5 +1,6 @@
-package lotto.utils;
+package lotto.generator;
 
+import lotto.generator.RandomLottoNumbersGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
